refactor: splite quiz Theory section in technicalInterview component to smaller and extract custom hooks
